UseScopeImpl
